% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm.R
\name{calcDistrParts}
\alias{calcDistrParts}
\title{Calculate Parts for Fisher Scoring}
\usage{
calcDistrParts(y, X, formula, data, w = NULL, params_char)
}
\arguments{
\item{y}{Response vector (binary)}

\item{X}{Data matrix}

\item{formula}{Formula used for the probit regression}

\item{data}{Data as string}

\item{w}{Weights}

\item{beta}{Current estimator}
}
\value{
List with XtX and Xy
}
\description{
This function calculates the parts required to conduct an update for the Fisher scoring for probit regression
}
\author{
Stefan B., Daniel S.
}
