% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm.R
\name{computePlacementValues}
\alias{computePlacementValues}
\title{Compute Placement Values on Server}
\usage{
computePlacementValues(score, truth, tset = NULL)
}
\arguments{
\item{score}{Prediction scores}

\item{truth}{True zero-one values as integer}

\item{tset}{Set of thresholds}
}
\value{
Placement values
}
\description{
This function calculates the placement values required to calculate the ROC-GLM.
}
\author{
Stefan B., Daniel S.
}
