% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm.R
\name{predictStringGLM}
\alias{predictStringGLM}
\title{Predict with given DataSHIELD GLM}
\usage{
predictStringGLM(params_char, data, formula, link = "logit")
}
\arguments{
\item{params_char}{Character containing all the parameter}

\item{data}{Character containing the name of the dataset}

\item{formula}{The formula used to fit the model}
}
\value{
Prediction scores
}
\description{
This function calculates the prediction scores for a given GLM
}
\author{
Stefan B., Daniel S.
}
