% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm.R
\name{rocGLMFram}
\alias{rocGLMFram}
\title{Calculate data for ROC-GLM}
\usage{
rocGLMFram(params_char, data, formula)
}
\arguments{
\item{params_char}{Character containing all the parameter}

\item{data}{Character containing the name of the dataset}

\item{formula}{The formula used to fit the model}
}
\value{
Data.frame used to calculate ROC-GLM
}
\description{
This function stores the data on the server used for the ROC-GLM#' @param params_char Character containing all the parameter
}
\author{
Stefan B., Daniel S.
}
